Image is being uploaded by the proxy
Configured a shared folder called statis on the yaml file in the root directly
tahtr connects the static folder in the flask aplication (backend service) and the
nginx proxy service.

Not sure how all this is working becasue i dont specify any shared folder on the
conf file insie the proxy folder.